src/github_api/views.py
from django.shortcuts import render
from github import Github, GithubException
import requests


def github(request):
    user = {}
    if 'username' in request.GET:
        username = request.GET['username']
        url = 'https://api.github.com/users/' + username
        response = requests.get(url)
        user = response.json()
    print(user)
    context = {
        'user': user
    }
    return render(request, 'github.html', context)


def github_client(request):
    search_result = {}
    if 'username' in request.GET:
        username = request.GET['username']
        client = Github()

        try:
            user = client.get_user(username)
            search_result['name'] = user.name
            search_result['login'] = user.login
            search_result['public_repos'] = user.public_repos
            search_result['success'] = True
        except GithubException as ge:
            search_result['message'] = get.data['message']
            search_result['success'] = False

        rate_limit = client.get_rate_limit()
        search_result['rate'] = {
            'limit': rate_limit.rate.limit,
            'remaining': rate_limit.rate.remaining,
        }

    context = {
        'search_result': search_result,
    }

    return render(request, 'github.html', context)
