requirements.txt
certifi==2019.9.11
chardet==3.0.4
Deprecated==1.2.6
Django==2.2
idna==2.8
PyGithub==1.44
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
wrapt==1.11.2
